ref(trace): remove scrolling from tree expand stuff
